Implement Launch object
This object models the launch of a Probe towards a particular target
area and can determine if the Probe will successfully enter the target
area or miss left or miss right.

Currently I put the logic for missing on the target area out of
expedience. I don't think that is right, and that logic should probably
live on Probe (or maybe in Launch itself)