Solve Day 21 Part Two
The classes and framework I made for part 1 weren't really a lot of help
with the dirac die so I just implemented the entire solution in a single
loop.

The key insight is that you don't have to model every game individually,
you can instead keep track of how many games have the same state and
increment accordingly.

This is pretty dense and has some duplicated code that could probably be
cleaned up but it works so I'm taking it.